@@ -1,29 +1,37 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
+apply plugin: 'com.google.gms.google-services'
 
 android {
-    compileSdkVersion 34
+    compileSdkVersion 36
+
     defaultConfig {
         applicationId "com.segmentify.segmentifysdk"
-        minSdkVersion 16
-        targetSdkVersion 34
+        minSdkVersion 24
+        targetSdkVersion 36
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
+
     buildFeatures {
         viewBinding = true
     }
+
     buildTypes {
         release {
             minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+
     namespace 'com.segmentify.segmentifysdk'
+
     lint {
         abortOnError false
     }
+
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_11
         targetCompatibility JavaVersion.VERSION_11
@@ -36,26 +44,32 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.google.firebase:firebase-core:16.0.1'
-    implementation 'com.google.firebase:firebase-messaging:17.3.0'
-    implementation 'com.google.firebase:firebase-iid:17.0.0'
+
+    implementation(platform("com.google.firebase:firebase-bom:34.0.0"))
+    implementation("com.google.firebase:firebase-messaging")
+    implementation("com.google.firebase:firebase-analytics")
+    implementation("com.google.firebase:firebase-firestore")
+
+    implementation("com.google.android.gms:play-services-auth:21.3.0")
+    implementation("androidx.work:work-runtime-ktx:2.10.2")
+    implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.9.1")
 
     implementation 'com.squareup.picasso:picasso:2.71828'
-    implementation 'com.scottyab:secure-preferences-lib:0.1.4'
-
+    implementation 'com.scottyab:secure-preferences-lib:0.1.7'
     implementation 'com.google.code.gson:gson:2.10.1'
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:${kotlin_version}"
-    implementation 'com.android.support:appcompat-v7:28.0.0'
 
-    implementation 'com.android.support:recyclerview-v7:28.0.0'
+    implementation("androidx.legacy:legacy-support-v4:1.0.0")
+    implementation "androidx.appcompat:appcompat:1.7.1"
+    implementation "androidx.recyclerview:recyclerview:1.4.0"
+    implementation "androidx.constraintlayout:constraintlayout:2.2.1"
+    implementation "androidx.constraintlayout:constraintlayout-compose:1.1.1"
 
-
-    implementation 'com.android.support.constraint:constraint-layout:2.0.4'
     testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test:runner:1.7.0'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.7.0'
+
     implementation project(path: ':segmentifyandroidsdk')
 }
-
 
 apply plugin: 'com.google.gms.google-services'
